# core/openapi.yaml
openapi: 3.0.0
info:
  title: WACAST Core API
  description: WhatsApp Multi-Session Gateway API - Enterprise WhatsApp messaging platform
  version: 1.0.0
  contact:
    name: WACAST Support
    email: support@wacast.io
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.wacast.io
    description: Production Server
...
paths:
...
  /api/v1/sessions/{device_id}:
    get:
      tags:
        - Sessions
      summary: Get session status
      description: Returns current status of a specific WhatsApp session
      operationId: getSessionStatus
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
          description: Device/Session identifier
          example: device001
      responses:
        "200":
          description: Session status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatusResponse"
        "404":
          description: Session not found
...
components:
  schemas:
...
    SessionStatusResponse:
      type: object
      properties:
        device_id:
          type: string
        status:
          type: integer
        is_active:
          type: boolean
        last_activity:
          type: integer
...